refactor: Rename dataset_list.py to list.py and move get_dataloader() & transform() to list.py
