+++ src/test.py
@@ -4,29 +4,12 @@ from pathlib import Path
 import torch
 from torch.utils.data import DataLoader
 
-from dataset.dataset_lits import LiTSSliceDataset
+from dataset.list import get_dataloader, transform
 from inference_executor import NestedUNetInferenceRunner
-from metrics import f1_score, f2_score, sensitivity, specificity
+from metrics import f1_score, f2_score, get_metric_name, sensitivity, specificity
 from model import NestedUNet
 
-
-def get_dataloader(
-    src_dir,
-    patient_ids,
-    batch_size,
-    cache_slice_info_path=None,
-):
-    dataset = LiTSSliceDataset(
-        src_dir=src_dir,
-        patient_ids=patient_ids,
-        cache_slice_info_path=cache_slice_info_path,
-    )
-    dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=4)
-    return dataloader
-
-
 if __name__ == "__main__":
-    # os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "expandable_segments:True"
     os.environ["CUDA_VISIBLE_DEVICES"] = "0"
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
@@ -53,6 +36,7 @@ if __name__ == "__main__":
         src_dir=src_dir,
         patient_ids=test_patient_ids,
         batch_size=batch_size,
+        transform=transform,
         cache_slice_info_path=test_slice_info_path,
     )
     if not test_slice_info_path.exists():
@@ -69,4 +53,4 @@ if __name__ == "__main__":
     )
 
     for idx, (metric_fn, metric) in enumerate(zip(metric_fns, metrics)):
-        print(f"{idx}. {metric_fn}: {metric:.2f}")
+        print(f"{idx}. {get_metric_name(metric_fn)}: {metric:.2f}")
